refactor: move ButtonBase to `_private` directory and changes Styleguidist ignore pattern to match
