+++ src/components/buttons/_private/ButtonBase/ButtonBase.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/_partials/index';
-@import '../../../common/styles/themeUtils';
+@import '../../../../styles/_partials/index';
+@import '../../../../common/styles/themeUtils';
 
 $disabled-background-light: $gray25;
 $disabled-background-dark: $gray;
