+++ src/components/buttons/_private/ButtonBase/ButtonBase.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import classnames from 'classnames';
 import * as styles from './ButtonBase.scss';
-import ILocalContainerProps from '../../../common/structures/ILocalContainerProps';
-import { Container } from '../../modules/Container';
+import ILocalContainerProps from '../../../../common/structures/ILocalContainerProps';
+import { Container } from '../../../modules/Container';
 
 export enum ButtonPropColor {
 	default = 'default',
